Pull out sendCreatePersonaMessage
